+++ CHANGELOG.md
@@ -17,8 +17,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - agent sync lower timeout
 - #8 perhaps i need to get wazuh setup to aquire scanning time
 - consider some wazuhs tab access profiles
-- pagination for alerts (from lasttime)
 - alerts deduplication
+- settings for plugin: refreshing time, request api fetch page size,
+
+## [0.0.6] 2025-04-07
+### Added
+- Wazuh Alerts
 
 ## [0.0.5] 2025-04-05
 ### Added